Multichannel SACD transcoding to avoid 5-channel flac
ALSA does not support odd channel numbers (any more) due to hardware
limitations (in some hardware).  So the player has to adapt the stream
to 6 channels and this is where that we start having inconsistencies.

A fail-safe solution is to add the missing LFE channel which is paired
with the centre channel, so transcoding both 5-ch and 6-ch SACD to
6-channel.  If the LFE channel is not present, we don't invent it but
insert an all-zero track.  Thanks to flac encoding, the all-zero track
contributes very little to the overall file size.